update package name to bagelML
