+++ docs/docs/integrations/providers/bagel.mdx
@@ -1,6 +1,6 @@
-# BagelDB
+# Bagel
 
-> [BagelDB](https://www.bageldb.ai/) (`Open Vector Database for AI`), is like GitHub for AI data.
+> [Bagel](https://www.bagel.net/) (`Open Vector Database for AI`), is like GitHub for AI data.
 It is a collaborative platform where users can create,
 share, and manage vector datasets. It can support private projects for independent developers,
 internal collaborations for enterprises, and public contributions for data DAOs.
@@ -8,13 +8,13 @@ internal collaborations for enterprises, and public contributions for data DAOs.
 ## Installation and Setup
 
 ```bash
-pip install betabageldb
+pip install bagelML
 ```
 
 
 ## VectorStore
 
-See a [usage example](/docs/integrations/vectorstores/bageldb).
+See a [usage example](/docs/integrations/vectorstores/bagel).
 
 ```python
 from langchain_community.vectorstores import Bagel
